Answer with sendonly when remote is recvonly
When receiving a remote offer with a transceiver which is recvonly, the
answer must not set the transceiver to a receiving mode to avoid Firefox
from complaining and failing to set the remote description with the
following error:

```
Answer tried to set recv when offer did not set send
```

While Chrome is more forgiving and happily sets remote transceiver which
include receive even when the offer did not request ist, this needs to
be changed to get interop with Firefox.

The change is to create the implicitly created receivers introduced with
PR#1179 as inactive if the remote description does offer only to
receive. This will later on switch the transceiver to sendonly as soon
as a matching track is added.